feat: add <AddPlayerToTeam/>
