+++ app/roster/page.tsx
@@ -52,6 +52,7 @@ import TeamBadgeStatus from "../components/team-badge-status";
 import GenderSwitch from "../components/genderSwitch";
 // --utils
 import { getMyCookie, hasMyCookie, setMyCookie } from "../utils/manage-cookies";
+import AddPlayerToTeam from "../components/add-player-to-team";
 
 const RosterPage = () => {
   const [hasTeam, setHasTeam] = useState<boolean>(false);
@@ -356,7 +357,7 @@ const RosterPage = () => {
       }));
     }
   };
-
+  //-- to set the team on cookie
   const handleBadgeClick = () => {
     const hasCookie = hasMyCookie("myTeam");
     if (hasCookie) return;
@@ -375,59 +376,10 @@ const RosterPage = () => {
       });
     }
   };
+
   const selectPlayer = async (player: Player) => {
     setSelectedPlayer(player);
     setIsDialogOpen(true);
-  };
-  const handleConfirmAddPlayer = async () => {
-    if (selectedPlayer && selectedTeam) {
-      setIsAdding(true);
-
-      try {
-        const response = await fetch("/api/addPlayerToTeam", {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({
-            player: selectedPlayer,
-            teamName: selectedTeam,
-            isOpponent: false,
-          }),
-        });
-
-        if (!response.ok) {
-          throw new Error("Failed to add player");
-        }
-
-        const data = await response.json();
-        setTeam(data.team);
-        toast({
-          title: "Player added",
-          description: `Player ${selectedPlayer.name} added to team ${selectedTeam}`,
-          variant: "default",
-          duration: 3000,
-        });
-      } catch (error) {
-        console.error("Error adding player:", error);
-        toast({
-          title: "Error",
-          description: "Failed to add player to team",
-          variant: "destructive",
-          duration: 3000,
-        });
-      } finally {
-        setIsAdding(false);
-        setIsDialogOpen(false);
-      }
-    } else {
-      toast({
-        title: "Selection required",
-        description: "Please select a player and a team before adding.",
-        variant: "destructive",
-        duration: 3000,
-      });
-    }
   };
 
   return (
@@ -575,176 +527,7 @@ const RosterPage = () => {
         </TabsContent>
 
         <TabsContent value="addPlayer">
-          <Card className="w-full mb-4">
-            <CardHeader>
-              <CardTitle>Add Player</CardTitle>
-            </CardHeader>
-            <CardContent>
-              <Input
-                type="text"
-                value={name}
-                onChange={(e) => setName(e.target.value)}
-                placeholder="Enter player's name"
-              />
-            </CardContent>
-            <CardFooter className="border-t px-6 py-4">
-              <Button onClick={handleSearch}>
-                {isSearching ? (
-                  <>
-                    <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-                    <Label>Please wait</Label>
-                  </>
-                ) : (
-                  "Search"
-                )}
-              </Button>
-            </CardFooter>
-          </Card>
-
-          <Card className="w-full">
-            <CardHeader>
-              <CardTitle>Search Results</CardTitle>
-            </CardHeader>
-            <CardContent>
-              <div className="overflow-x-auto">
-                <Table>
-                  <TableHeader>
-                    <TableRow>
-                      <TableHead className="whitespace-nowrap">Name</TableHead>
-                      <TableHead className="whitespace-nowrap">
-                        Rating
-                      </TableHead>
-                      <TableHead className="whitespace-nowrap">
-                        PDGA Number
-                      </TableHead>
-                      <TableHead className="whitespace-nowrap">Class</TableHead>
-                      <TableHead className="whitespace-nowrap">City</TableHead>
-                      <TableHead className="whitespace-nowrap">State</TableHead>
-                      <TableHead className="whitespace-nowrap">
-                        Country
-                      </TableHead>
-                      <TableHead className="whitespace-nowrap">
-                        Membership Status
-                      </TableHead>
-                      <TableHead className="whitespace-nowrap">
-                        Action
-                      </TableHead>
-                    </TableRow>
-                  </TableHeader>
-                  {results.length === 0 ? (
-                    <TableBody>
-                      <TableRow>
-                        <TableCell colSpan={9} className="text-center pt-10">
-                          {isSearching ? (
-                            <Loader2
-                              size="32"
-                              className="mx-auto animate-spin"
-                            />
-                          ) : (
-                            <Label className="text-sm">No results found</Label>
-                          )}
-                        </TableCell>
-                      </TableRow>
-                    </TableBody>
-                  ) : (
-                    <TableBody>
-                      {results.map((player, index) => (
-                        <TableRow key={index}>
-                          <TableCell className="whitespace-nowrap">
-                            {player.name}
-                          </TableCell>
-                          <TableCell className="whitespace-nowrap">
-                            {player.rating}
-                          </TableCell>
-                          <TableCell className="whitespace-nowrap">
-                            {player.pdgaNumber}
-                          </TableCell>
-                          <TableCell className="whitespace-nowrap">
-                            {player.class}
-                          </TableCell>
-                          <TableCell className="whitespace-nowrap">
-                            {player.city}
-                          </TableCell>
-                          <TableCell className="whitespace-nowrap">
-                            {player.state}
-                          </TableCell>
-                          <TableCell className="whitespace-nowrap">
-                            {player.country}
-                          </TableCell>
-                          <TableCell className="whitespace-nowrap">
-                            {player.membershipStatus}
-                          </TableCell>
-                          <TableCell>
-                            <Button
-                              onClick={() => selectPlayer(player)}
-                              disabled={isAdding}
-                              className="whitespace-nowrap"
-                            >
-                              {isAdding ? (
-                                <Loader2 className="h-4 w-4 animate-spin" />
-                              ) : (
-                                "Select"
-                              )}
-                            </Button>
-                          </TableCell>
-                        </TableRow>
-                      ))}
-                    </TableBody>
-                  )}
-                </Table>
-                <Dialog open={isDialogOpen} onOpenChange={setIsDialogOpen}>
-                  <DialogContent className="w-[90%] md:w-[100%] lg:w-[100%] rounded-lg">
-                    <DialogHeader className="text-start">
-                      <DialogTitle>Add Player to Team</DialogTitle>
-                      <DialogDescription className="!mt-1">
-                        Select gender and team for {selectedPlayer?.name}
-                      </DialogDescription>
-                    </DialogHeader>
-
-                    <div className="flex gap-4 mb-0 md:mb-2 lg:mb-2 pt-2 items-center">
-                      <Label>Gender</Label>
-                      <GenderSwitch
-                        selectedGender={selectedGender}
-                        onGenderChange={setSelectedGender}
-                      />
-                    </div>
-
-                    <div className="flex flex-row gap-2 items-center mb-2 md:mb-0 lg:mb-0">
-                      <Label>Team</Label>
-                      <DropdownMenu>
-                        <DropdownMenuTrigger asChild>
-                          <Button variant="outline">
-                            {selectedTeam ? selectedTeam : "Select a team"}
-                          </Button>
-                        </DropdownMenuTrigger>
-                        <DropdownMenuContent className="custom-dropdown-content overflow-y-auto max-h-[250px]">
-                          {teamNames.map((team) => (
-                            <DropdownMenuItem
-                              key={team}
-                              onSelect={() => setSelectedTeam(team)}
-                              className="w-full py-2 px-4 hover:bg-gray-100 pl-7"
-                            >
-                              {team}
-                            </DropdownMenuItem>
-                          ))}
-                        </DropdownMenuContent>
-                      </DropdownMenu>
-                    </div>
-
-                    <DialogFooter>
-                      <Button onClick={handleConfirmAddPlayer}>
-                        {isAdding ? (
-                          <Loader2 className="h-4 w-4 animate-spin" />
-                        ) : (
-                          "Confirm"
-                        )}
-                      </Button>
-                    </DialogFooter>
-                  </DialogContent>
-                </Dialog>
-              </div>
-            </CardContent>
-          </Card>
+          <AddPlayerToTeam />
         </TabsContent>
       </Tabs>
     </div>
